server/requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
alembic==1.14.1
altair==5.5.0
annotated-types==0.7.0
anyio==4.8.0
appdirs==1.4.4
asgiref==3.8.1
asn1crypto==1.5.1
asttokens @ file:///opt/conda/conda-bld/asttokens_1646925590279/work
atpublic==5.0
attrs==24.3.0
auth0-python==4.7.2
Authlib==1.3.1
backoff==2.2.1
bcrypt==4.2.1
beautifulsoup4==4.12.3
blinker==1.9.0
build==1.2.2.post1
cachetools==5.5.0
certifi==2024.12.14
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
chroma-hnswlib==0.7.6
chromadb==0.5.23
click==8.1.8
cohere==5.13.11
colorama @ file:///C:/ci_311/colorama_1676422310965/work
coloredlogs==15.0.1
comm @ file:///C:/b/abs_67a8058udb/croot/comm_1709322909844/work
contourpy==1.3.1
crewai==0.98.0
crewai-tools==0.32.1
cryptography==43.0.3
cycler==0.12.1
dataclasses-json==0.6.7
datasets==3.2.0
debugpy @ file:///C:/b/abs_bf9oo2vhxp/croot/debugpy_1736269476451/work
decorator @ file:///opt/conda/conda-bld/decorator_1643638310831/work
Deprecated==1.2.15
deprecation==2.1.0
dill==0.3.8
distro==1.9.0
dnspython==2.7.0
docker==7.1.0
docstring_parser==0.16
docx2txt==0.8
duckduckgo_search==7.3.0
durationpy==0.9
embedchain==0.1.126
et_xmlfile==2.0.0
executing @ file:///opt/conda/conda-bld/executing_1646925071911/work
fastapi==0.115.6
fastavro==1.10.0
filelock==3.16.1
filetype==1.2.0
flatbuffers==24.12.23
fonttools==4.55.6
frozendict==2.4.6
frozenlist==1.5.0
fsspec==2024.9.0
gitdb==4.0.12
GitPython==3.1.44
google-ai-generativelanguage==0.6.10
google-api-core==2.24.0
google-api-python-client==2.158.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-cloud-aiplatform==1.78.0
google-cloud-bigquery==3.29.0
google-cloud-core==2.4.1
google-cloud-resource-manager==1.14.0
google-cloud-storage==2.19.0
google-crc32c==1.6.0
google-generativeai==0.8.3
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
gptcache==0.1.44
greenlet==3.1.1
grpc-google-iam-v1==0.14.0
grpcio==1.69.0
grpcio-health-checking==1.69.0
grpcio-status==1.69.0
grpcio-tools==1.69.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.7
httplib2==0.22.0
httptools==0.6.4
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.27.1
humanfriendly==10.0
hyperframe==6.0.1
idna==3.10
ijson==3.3.0
importlib_metadata==8.5.0
importlib_resources==6.5.2
instructor==1.7.2
ipykernel @ file:///C:/b/abs_5etv6sld_j/croot/ipykernel_1728665608850/work
ipython @ file:///C:/b/abs_8eyhzleyrk/croot/ipython_1734548134403/work
jedi @ file:///C:/b/abs_3a2kbnlclc/croot/jedi_1733987412687/work
Jinja2==3.1.5
jiter==0.8.2
joblib==1.4.2
json_repair==0.35.0
jsonpatch==1.33
jsonpickle==4.0.1
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client @ file:///C:/b/abs_a6h3c8hfdq/croot/jupyter_client_1699455939372/work
jupyter_core @ file:///C:/b/abs_beftpbuevw/croot/jupyter_core_1718818307097/work
keybert==0.8.5
kiwisolver==1.4.8
kubernetes==31.0.0
lancedb==0.18.0
langchain==0.3.14
langchain-chroma==0.2.0
langchain-cohere==0.3.5
langchain-community==0.3.14
langchain-core==0.3.29
langchain-experimental==0.3.4
langchain-google-community==2.0.4
langchain-google-genai==2.0.8
langchain-mongodb==0.4.0
langchain-openai==0.2.14
langchain-text-splitters==0.3.5
langgraph==0.2.64
langgraph-checkpoint==2.0.10
langgraph-sdk==0.1.51
langsmith==0.1.147
linkup-sdk==0.2.2
litellm==1.57.4
lxml==5.3.0
Mako==1.3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.25.1
matplotlib==3.10.0
matplotlib-inline @ file:///C:/ci_311/matplotlib-inline_1676425798036/work
mdurl==0.1.2
mem0ai==0.1.47
mmh3==5.0.1
model2vec==0.3.7
monotonic==1.6
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
multiprocess==0.70.16
mypy-extensions==1.0.0
narwhals==1.21.1
nest-asyncio @ file:///C:/b/abs_65d6lblmoi/croot/nest-asyncio_1708532721305/work
networkx==3.4.2
nodeenv==1.9.1
numpy==1.26.4
oauthlib==3.2.2
onnxruntime==1.20.1
openai==1.60.0
openpyxl==3.1.5
opentelemetry-api==1.29.0
opentelemetry-exporter-otlp-proto-common==1.29.0
opentelemetry-exporter-otlp-proto-grpc==1.29.0
opentelemetry-exporter-otlp-proto-http==1.29.0
opentelemetry-instrumentation==0.50b0
opentelemetry-instrumentation-asgi==0.50b0
opentelemetry-instrumentation-fastapi==0.50b0
opentelemetry-proto==1.29.0
opentelemetry-sdk==1.29.0
opentelemetry-semantic-conventions==0.50b0
opentelemetry-util-http==0.50b0
orjson==3.10.14
outcome==1.3.0.post0
overrides==7.7.0
packaging @ file:///C:/b/abs_3by6s2fa66/croot/packaging_1734472138782/work
pandas==2.2.3
parso @ file:///C:/b/abs_834b4mj92b/croot/parso_1733963322289/work
patronus==0.0.17
pdfminer.six==20231228
pdfplumber==0.11.5
pillow==11.1.0
platformdirs @ file:///C:/b/abs_b6z_yqw_ii/croot/platformdirs_1692205479426/work
portalocker==2.10.1
posthog==3.7.5
primp==0.11.0
prompt-toolkit @ file:///C:/b/abs_68uwr58ed1/croot/prompt-toolkit_1704404394082/work
propcache==0.2.1
proto-plus==1.25.0
protobuf==5.29.3
psutil @ file:///C:/b/abs_b5gv3mn55h/croot/psutil_1736371546320/work
psycopg2==2.9.10
pure-eval @ file:///opt/conda/conda-bld/pure_eval_1646925070566/work
pyarrow==18.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.5
pydantic-settings==2.7.1
pydantic_core==2.27.2
pydeck==0.9.1
Pygments @ file:///C:/b/abs_fay9dpq4n_/croot/pygments_1684279990574/work
PyJWT==2.10.1
pylance==0.22.0
pymongo==4.10.1
pyOpenSSL==24.3.0
pyparsing==3.2.1
pypdf==5.1.0
pypdfium2==4.30.1
PyPika==0.48.9
pyproject_hooks==1.2.0
pyreadline3==3.5.4
pyright==1.1.392.post0
pysbd==0.3.4
PySocks==1.7.1
python-dateutil @ file:///C:/b/abs_3au_koqnbs/croot/python-dateutil_1716495777160/work
python-dotenv==1.0.1
pytube==15.0.0
pytz==2024.2
pyvis==0.3.2
pywin32==308
PyYAML==6.0.2
pyzmq @ file:///C:/b/abs_f3yte6j5yn/croot/pyzmq_1734711069724/work
qdrant-client==1.13.2
rank-bm25==0.2.2
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.22.3
rsa==4.9
safetensors==0.5.2
schema==0.7.7
scikit-learn==1.6.1
scipy==1.15.1
scrapegraph_py==1.10.2
seaborn==0.13.2
selenium==4.28.0
semhash==0.2.0
sentence-transformers==3.3.1
serpapi==0.1.5
shapely==2.0.6
shellingham==1.5.4
simhash==2.1.2
six @ file:///tmp/build/80754af9/six_1644875935023/work
smmap==5.0.2
sniffio==1.3.1
snowflake==1.0.2
snowflake-connector-python==3.12.4
snowflake._legacy==1.0.0
snowflake.core==1.0.2
sortedcontainers==2.4.0
soupsieve==2.6
spider-client==0.1.26
SQLAlchemy==2.0.37
stack-data @ file:///opt/conda/conda-bld/stack_data_1646927590127/work
starlette==0.41.3
streamlit==1.41.1
sympy==1.13.1
tabulate==0.9.0
tenacity==9.0.0
threadpoolctl==3.5.0
tiktoken==0.7.0
tokenizers==0.21.0
toml==0.10.2
tomli==2.2.1
tomli_w==1.2.0
tomlkit==0.13.2
torch==2.6.0
tornado @ file:///C:/b/abs_7cyu943ybx/croot/tornado_1733960510898/work
tqdm==4.67.1
traitlets @ file:///C:/b/abs_bfsnoxl4pq/croot/traitlets_1718227069245/work
transformers==4.48.0
trio==0.28.0
trio-websocket==0.11.1
typer==0.15.1
types-requests==2.32.0.20241016
typing-inspect==0.9.0
typing_extensions @ file:///C:/b/abs_0ffjxtihug/croot/typing_extensions_1734714875646/work
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.3.0
usearch==2.16.9
uv==0.5.22
uvicorn==0.34.0
validators==0.34.0
vicinity==0.4.0
watchdog==6.0.0
watchfiles==1.0.4
wcwidth @ file:///Users/ktietz/demo/mc3/conda-bld/wcwidth_1629357192024/work
weaviate-client==4.10.4
websocket-client==1.8.0
websockets==14.1
wrapt==1.17.1
wsproto==1.2.0
xxhash==3.5.0
yarl==1.18.3
zipp==3.21.0
